+++ src/config.ts
@@ -121,10 +121,6 @@ export interface IConfig {
       serverNames: string[];
       banMessage: string;
     };
-    mentionLimitProtection: {
-      maxMentions: number;
-      redactReason: string;
-    };
   };
   safeMode?: {
     bootOption: SafeModeBootOption;
@@ -228,11 +224,6 @@ const defaultConfig: IConfig = {
       serverNames: [],
       banMessage:
         "Unfortunately we cannot accept new users from your homeserver at this time.",
-    },
-    mentionLimitProtection: {
-      maxMentions: 3,
-      redactReason:
-        "You have mentioned too many users in this message, so we have had to redact it.",
     },
   },
   safeMode: {
